requirements.txt
csscompressor==0.9.4
Django==1.10.5
django-analytical==2.2.2
django-appconf==1.0.2
django-bower==5.2.0
django-compressor==2.1
django-debug-toolbar==1.6
django-jet==1.0.4
django-materializecss-form==1.0.1
django-robots==2.0
django-sass-processor==0.5.3
django-seo2==1.0.1
libsass==0.12.3
psycopg2==2.6.2
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.2
